V1.0
Implement improved drone routing algorithm with cost-aware pathfinding

- Enhanced pathfinding to track total path cost instead of greedy heuristic
- Added cost calculation for proper restricted zone handling (cost=2.0)
- Improved drone sequencing for better throughput through bottleneck zones
- Fixed restricted zone transit handling with -- notation for in-transit drones
- Code cleanup and refactoring for better maintainability

+++ models/zone.py
@@ -8,7 +8,7 @@ class Zone:
         self.color: str = 'none'
         self.is_start: bool = False
         self.is_end: bool = False
-        self.cost: int = 1
+        self.cost: float = 1.0
         self.reserved: int = 0
         self.capacity: int = 0
